requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.4.58
Pillow==8.4.0
PyYAML==6.0
requests==2.26.0
scipy==1.5.4  # Google Colab version
torch==1.8.0
torchvision==0.9.0
tqdm==4.62.3
protobuf==3.19.1  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard==1.15.0

# Plotting ------------------------------------
pandas==1.1.5
seaborn==0.11.2

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation

# Web-API --------------------------------
flask
flask-cors
gunicorn

# ocr
tensorflow==1.15.0


